Refactor course_image_url
