+++ cms/djangoapps/models/settings/course_details.py
@@ -8,9 +8,10 @@ from django.conf import settings
 
 from opaque_keys.edx.locations import Location
 from xmodule.modulestore.exceptions import ItemNotFoundError
-from contentstore.utils import course_image_url, has_active_web_certificate
+from contentstore.utils import has_active_web_certificate
 from models.settings import course_grading
 from openedx.core.djangoapps.self_paced.models import SelfPacedConfiguration
+from openedx.core.lib.courses import course_image_url
 from xmodule.fields import Date
 from xmodule.modulestore.django import modulestore
