+++ lms/djangoapps/course_structure_api/v0/serializers.py
@@ -3,7 +3,7 @@
 from django.core.urlresolvers import reverse
 from rest_framework import serializers
 
-from courseware.courses import course_image_url
+from openedx.core.lib.courses import course_image_url
 
 
 class CourseSerializer(serializers.Serializer):
