+++ lms/templates/shoppingcart/receipt.html
@@ -3,9 +3,10 @@
 from django.utils.translation import ugettext as _
 from django.utils.translation import ungettext
 from django.core.urlresolvers import reverse
-from courseware.courses import course_image_url, get_course_about_section, get_course_by_id
+from courseware.courses import get_course_about_section, get_course_by_id
 from markupsafe import escape
 from microsite_configuration import microsite
+from openedx.core.lib.courses import course_image_url
 %>
 
 <%block name="billing_details_highlight">
