+++ lms/templates/shoppingcart/registration_code_redemption.html
@@ -1,7 +1,8 @@
 <%!
 from django.utils.translation import ugettext as _
 from django.core.urlresolvers import reverse
-from courseware.courses import course_image_url, get_course_about_section
+from courseware.courses import get_course_about_section
+from openedx.core.lib.courses import course_image_url
 %>
 <%inherit file="../main.html" />
 <%namespace name='static' file='/static_content.html'/>
